[front] re-work how new positions are found for moving tasks
The old method wasn't just unreliable, it was just plain wrong :D

the old method counted how many tasks were in a column/swimlane combination and used it's index as position. This is wrong on many levels, as not all tasks are visible so counting them is just incorrect.

The new method specifically sets the tasks position as a data attribute which is then retrieved onDragover and stored in state until dropped. it is assumed that if there is no state then it has been dropped on a column with no tasks and uses index 0 as the new position.