[server][taskService] set task.column and task.swimlane to empty models before updating
I have added a comment within the code here to come back and find out why I have had to do this.

For context:

Before this change the column and swimlane would not be updated when using the generic repository. After some debugging I found that the code was correctly requesting the taskrepository to update the task.columnid and task.swimlaneid, however at the point of update the task still had an old preloaded task.swimlane and task.column which depending on the update may not have matched the new id's. After setting the column  and swimlane  as empty structs this behaves as expected.

It is important to come back to this (thus the code comment) as if I do not understand this behavior in Gorm it is bound to cause more issues