refactor(database): server: move Migrate() to the generic repository definition
Originally there were plans to add custom options to automigrate per-model (e.g. if the db was clickhouse we may order columns on disk in an optimised manor) however as we aren't currently using this and it requires so much duplicated code for every new model we are going to remove this option to cleanup the codebase.

This will likely be reviewed in the future. However this will bring more things into question, such as whether to write specific migrations instead of relying on automigrate to decide data types on our behalf.